Move trait selection error reporting to its own top-level module
