+++ compiler/rustc_trait_selection/src/error_reporting/mod.rs
@@ -1,0 +1,1 @@
+pub mod traits;
