+++ compiler/rustc_trait_selection/src/traits/fulfill.rs
@@ -1,5 +1,5 @@
+use crate::error_reporting::traits::TypeErrCtxtExt;
 use crate::infer::{InferCtxt, TyOrConstInferVar};
-use crate::traits::error_reporting::TypeErrCtxtExt;
 use crate::traits::normalize::normalize_with_depth_to;
 use rustc_data_structures::captures::Captures;
 use rustc_data_structures::obligation_forest::ProcessResult;
